Resolving assorted warnings.

+++ lib/generator_function/generator_random_function.dart
@@ -3,9 +3,12 @@ part of graph.generatorfunction;
 /**
  * @author Mate
  * A generator random cost function
- * It will generate random (type "double") edge weights in the range of (<b>minWeight</b>, <b>maxWeight</b>) and rounds the values to <b>roundToDecimals</b>
+ * It will generate random (type "double") edge weights in the range of
+ * (<b>minWeight</b>, <b>maxWeight</b>) and rounds the values to
+ * <b>roundToDecimals</b>
  */
 class GeneratorRandomFunction extends GeneratorFunction {
+  
   double _maxWeight = 1.0;
 
   double _minWeight = 0.0;
@@ -21,7 +24,7 @@ class GeneratorRandomFunction extends GeneratorFunction {
     double edgeWeight = null;
 
     edgeWeight = math.random() * (_maxWeight - _minWeight) + _minWeight;
-    edgeWeight = math.round(edgeWeight * Math.pow(10, getRoundToDecimals())) / Math.pow(10, getRoundToDecimals()) as double;
+    edgeWeight = math.round(edgeWeight * Math.pow(10, getRoundToDecimals())) / Math.pow(10, getRoundToDecimals());
 
     return edgeWeight;
   }
@@ -46,4 +49,5 @@ class GeneratorRandomFunction extends GeneratorFunction {
   void setRoundToDecimals(int roundToDecimals) {
     this._roundToDecimals = roundToDecimals;
   }
+  
 }
